[fix] : moved MMatrix.h to MMatrix-class.h and changed all include minor updates on some doc NO WARNING ON R CMD check !! (with inconsolata, check --as-cran still TODO !)
