+++ R/method-mmatrix-add.descriptor.file.r
@@ -3,7 +3,7 @@
 #' @rdname descriptor
 #' @param object a mmatrix object
 #' 
-#' @details Creates a descriptor file, for reading the matrix with the package bigmemory.
+#' @details Creates a descriptor file, for reading the matrix with the package bigmemory or the \link{read.descriptor} function.
 #'
 #' @export
 setGeneric("add.descriptor.file", function(object) standardGeneric("add.descriptor.file"))
