--- man/add.descriptor.file.Rd
@@ -1,24 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/method-mmatrix-add.descriptor.file.r
-\name{add.descriptor.file}
-\alias{add.descriptor.file}
-\title{Add a Descriptor File to an mmatrix Object}
-\usage{
-add.descriptor.file(object)
-}
-\arguments{
-\item{object}{An object of class `mmatrix`. Must contain the slots `file`, `dim`, and `datatype`.}
-}
-\value{
-Invisibly returns `NULL`. The function is called for its side effect of writing a descriptor file.
-}
-\description{
-This generic function adds a descriptor file to an `mmatrix` object.
-}
-\details{
-This function wraps around an internal helper `mk.descriptor.file()` that creates a descriptor
-file for a matrix stored on disk.
-}
-\seealso{
-\link{mk.descriptor.file}
-}
