+++ man/mmatrix-class.Rd
@@ -6,7 +6,6 @@
 \title{Class \code{"mmatrix"}}
 \description{
 S4 class for manipulating memory-mapped files as matrices
-
 }
 \section{Slots}{
 
@@ -19,8 +18,7 @@ S4 class for manipulating memory-mapped files as matrices
 
 \item{\code{datatype}}{\code{character} giving the C++ underlying datatype.}
 
-\item{\code{readonly}}{\code{logical} Indicates if the array if read-only.}
-
+\item{\code{readonly}}{\code{logical} Indicates if the array is read-only.}
 }}
 
 \section{Objects from the Class}{
